Pull Request: Merge v1.3.x into master

+++ gators/feature_generation/tests/test_math_features.py
@@ -11,12 +11,12 @@ def test_transform_default():
     X = pl.DataFrame({"col1": [1, 2, 3], "col2": [4, 5, 6], "col3": [7, 8, 9]})
     expected_X = X.with_columns(
         [
-            (pl.col("col1") + pl.col("col2")).alias("col1_col2_sum"),
-            (pl.col("col1") + pl.col("col2") + pl.col("col3")).alias("col1_col2_col3_sum"),
+            (pl.col("col1") + pl.col("col2")).cast(pl.Float64).alias("col1_col2_sum"),
+            (pl.col("col1") + pl.col("col2") + pl.col("col3")).cast(pl.Float64).alias("col1_col2_col3_sum"),
             ((pl.col("col1") + pl.col("col2")) / 2.0).alias("col1_col2_mean"),
-            pl.min_horizontal(pl.col("col1"), pl.col("col2")).alias("col1_col2_min"),
+            pl.min_horizontal(pl.col("col1"), pl.col("col2")).cast(pl.Float64).alias("col1_col2_min"),
             ((pl.col("col1") + pl.col("col2") + pl.col("col3")) / 3.0).alias("col1_col2_col3_mean"),
-            pl.min_horizontal(pl.col("col1"), pl.col("col2"), pl.col("col3")).alias(
+            pl.min_horizontal(pl.col("col1"), pl.col("col2"), pl.col("col3")).cast(pl.Float64).alias(
                 "col1_col2_col3_min"
             ),
         ]
@@ -36,12 +36,12 @@ def test_transform_with_new_column_names():
     X = pl.DataFrame({"col1": [1, 2, 3], "col2": [4, 5, 6], "col3": [7, 8, 9]})
     expected_X = X.with_columns(
         [
-            (pl.col("col1") + pl.col("col2")).alias("12_sum"),
-            (pl.col("col1") + pl.col("col2") + pl.col("col3")).alias("123_sum"),
+            (pl.col("col1") + pl.col("col2")).cast(pl.Float64).alias("12_sum"),
+            (pl.col("col1") + pl.col("col2") + pl.col("col3")).cast(pl.Float64).alias("123_sum"),
             ((pl.col("col1") + pl.col("col2")) / 2.0).alias("12_mean"),
-            pl.min_horizontal(pl.col("col1"), pl.col("col2")).alias("12_min"),
+            pl.min_horizontal(pl.col("col1"), pl.col("col2")).cast(pl.Float64).alias("12_min"),
             ((pl.col("col1") + pl.col("col2") + pl.col("col3")) / 3.0).alias("123_mean"),
-            pl.min_horizontal(pl.col("col1"), pl.col("col2"), pl.col("col3")).alias("123_min"),
+            pl.min_horizontal(pl.col("col1"), pl.col("col2"), pl.col("col3")).cast(pl.Float64).alias("123_min"),
         ]
     )
 
@@ -61,12 +61,12 @@ def test_transform_with_drop_columns():
 
     expected_X = X.with_columns(
         [
-            (pl.col("col1") + pl.col("col2")).alias("col1_col2_sum"),
-            (pl.col("col1") + pl.col("col2") + pl.col("col3")).alias("col1_col2_col3_sum"),
+            (pl.col("col1") + pl.col("col2")).cast(pl.Float64).alias("col1_col2_sum"),
+            (pl.col("col1") + pl.col("col2") + pl.col("col3")).cast(pl.Float64).alias("col1_col2_col3_sum"),
             ((pl.col("col1") + pl.col("col2")) / 2.0).alias("col1_col2_mean"),
-            pl.min_horizontal(pl.col("col1"), pl.col("col2")).alias("col1_col2_min"),
+            pl.min_horizontal(pl.col("col1"), pl.col("col2")).cast(pl.Float64).alias("col1_col2_min"),
             ((pl.col("col1") + pl.col("col2") + pl.col("col3")) / 3.0).alias("col1_col2_col3_mean"),
-            pl.min_horizontal(pl.col("col1"), pl.col("col2"), pl.col("col3")).alias(
+            pl.min_horizontal(pl.col("col1"), pl.col("col2"), pl.col("col3")).cast(pl.Float64).alias(
                 "col1_col2_col3_min"
             ),
         ]
